// src/app/models/assignment.ts
export interface Assignment {
    id?: number;
    employeeId?: number;
    employeeName?: string;
    taskId?: number;
    taskName?: string;
    startTime?: string;
    dueTime?: string;
    completeTime?: string;
    managerId?: number;
    managerName?: string;
    statusId?: number;
    statusName?: string;
}